requirements.txt
alembic==1.5.8
amqp==5.0.6
aniso8601==9.0.1
appdirs==1.4.4
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
blinker==1.4
bootstrap4==0.1.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cssmin==0.2.0
defusedxml==0.7.1
distlib==0.3.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Faker==0.8.13
filelock==3.0.12
Flask==2.0.0
Flask-Admin==1.5.8
Flask-Assets==2.0
Flask-Babel==0.11.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.10.1
Flask-Celery-Helper==1.1.0
Flask-Dance==4.0.0
Flask-DebugToolbar==0.11.0
Flask-FontAwesome==0.1.5
Flask-JWT-Extended==4.3.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-OpenID==1.2.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.0
Flask-WTF==1.0.0
greenlet==1.0.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.0
jsmin==2.2.2
kombu==5.0.2
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
oauthlib==3.1.0
passlib==1.7.4
phonenumbers==8.12.45
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
speaklater==1.3
SQLAlchemy==1.4.11
text-unidecode==1.2
urllib3==1.26.4
URLObject==2.4.3
vine==5.0.0
virtualenv==20.4.2
visitor==0.1.3
wcwidth==0.2.5
webassets==2.0
Werkzeug==2.0.2
win32-setctime==1.0.3
WTForms==2.3.3
